requirements.txt
Django>=4.2
djangorestframework
django-tailwind[reload]
django-cors-headers
psycopg2-binary
python-decouple
dj-database-url
requests
pillow
gunicorn
whitenoise
